64bit 에서 컴파일할 때에 warning 이 없어지도록 수정함

+++ trunk/SipPlatform/StringMap.cpp
@@ -87,7 +87,7 @@ int CStringMap::GetCount( )
 	int iCount;
 
 	m_clsMutex.acquire();
-	iCount = m_clsMap.size();
+	iCount = (int)m_clsMap.size();
 	m_clsMutex.release();
 
 	return iCount;
